websocket/python/requirements.txt
websockets>=13
sounddevice
python-dotenv
aiohttp
